+++ src/index.ts
@@ -1,0 +1,5 @@
+export * from './datePlus'
+export * from './toDuration'
+export * from './toMs'
+export * from './toSeconds'
+
