+++ src/lib/divideDuration.ts.bak
@@ -1,0 +1,12 @@
+import { toSeconds } from '../toSeconds'
+
+// HELPER FUNCTION: creates convenience methods below
+export const divideDuration = (duration: string) =>
+  ({
+      by: (divisor: string) => {
+                                  const diff = toSeconds(duration)
+                                  const diffBy = toSeconds(divisor)
+
+                                  return diff / diffBy
+                                }
+  })
