+++ src/lib/units.ts
@@ -1,0 +1,9 @@
+export const units: Record<string, number> = {
+  'year': 365 * 24 * 60 * 60,
+  'month': 30 * 24 * 60 * 60,
+  'week': 7 * 24 * 60 * 60,
+  'day': 24 * 60 * 60,
+  'hour': 60 * 60,
+  'minute': 60,
+  'second': 1,
+}
